move constant to const.py
